+++ src/routes/+page.svelte
@@ -2,11 +2,11 @@
 
 <script lang="ts">
     import { ZPVDs, ZTVDs } from "$lib/db";
-    import ZakAvailable from "$lib/ZakAvailable.svelte";
     import { onMount } from "svelte";
     import Link from "$lib/Link.svelte";
     import RandomVideoFeed from "$lib/VideoTuple.svelte";
     import Bola from "$lib/Bola.svelte";
+    import ZakAvailable from "$lib/ZakAvailable.svelte";
 
     // eslint-disable no-unassigned-vars (sí se asigna con un bind: pero no me lo detecta)
     let container: HTMLDivElement;